chore: post-rebase cleaning

+++ packages/react-native-executorch/common/rnexecutorch/models/speech_to_text/types/Segment.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include <vector>
 #include "Word.h"
+#include <vector>
 
 namespace rnexecutorch::models::speech_to_text::types {
 
 struct Segment {
   std::vector<Word> words;
-  std::vector<int32_t> tokens; // Raw token IDs
+  std::vector<uint64_t> tokens; // Raw token IDs
   float start;
   float end;
   float avgLogprob;
